% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data}
\alias{data}
\alias{SE}
\title{Example dataset}
\value{
a \code{\link[SummarizedExperiment]{SummarizedExperiment-class}}.
}
\description{
A \code{\link[SummarizedExperiment]{SummarizedExperiment-class}} containing (a
subset of) hippocampus RNAseq of mice treated with Forskolin.
}
\references{
Chen et al. 2017.
Mapping Gene Expression in Excitatory Neurons during Hippocampal Late-Phase
Long-Term Potentiation
\emph{Frontiers in Molecular Neuroscience}.
DOI: 10.3389/fnmol.2017.00039
}
