% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{homogenizeDEA}
\alias{homogenizeDEA}
\title{homogenizeDEA}
\usage{
homogenizeDEA(x)
}
\arguments{
\item{x}{A data.frame containing the results of a differential expression
analysis}
}
\value{
A standardized data.frame.
}
\description{
Standardizes the outputs of differential expression methods (to an
  edgeR-like style)
}
