% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{qualitativeColors}
\alias{qualitativeColors}
\title{qualitativeColors}
\usage{
qualitativeColors(names, ...)
}
\arguments{
\item{names}{The names to which the colors are to be assigned, or an integer
indicating the desired number of colors}

\item{...}{passed to `randomcoloR::distinctColorPalette`}
}
\value{
A vector (eventually named) of colors
}
\description{
qualitativeColors
}
