% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{resetAllSechmOptions}
\alias{resetAllSechmOptions}
\title{resetAllSechmOptions}
\usage{
resetAllSechmOptions()
}
\value{
None
}
\description{
Resents all global options relative to sechm
}
\examples{
resetAllSechmOptions()

}
