% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{safescale}
\alias{safescale}
\title{safescale}
\usage{
safescale(x, center = TRUE, byRow = FALSE)
}
\arguments{
\item{x}{A matrix.}

\item{center}{Logical, whether to center values.}

\item{byRow}{Logical, whether to scale by rows instead of columns.}
}
\value{
A scaled matrix.
}
\description{
Equivalent to `base::scale`, but handling missing values and null variance
a bit more elegantly.
}
\examples{
m <- matrix(rnorm(100), nrow=10)
m.scaled <- safescale(m)
}
