% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{scale2}
\alias{scale2}
\title{scale2}
\usage{
scale2(x)
}
\arguments{
\item{x}{A matrix whose rows are to be scaled.}
}
\value{
A matrix of dimensions like x
}
\description{
A wrapper for non-centered unit-variance scaling
}
\examples{
scale2(matrix(1:9,nrow=3))
}
