% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{setSechmOption}
\alias{setSechmOption}
\title{setSechmOption}
\usage{
setSechmOption(variable, value)
}
\value{
None
}
\description{
Resents all global options relative to sechm
}
\examples{
setSechmOption("hmcols", value=c("blue","black","yellow"))

}
