% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{resetAllSechmOptions}
\alias{resetAllSechmOptions}
\title{resetAllSechmOptions}
\usage{
resetAllSechmOptions()
}
\value{
None
}
\description{
Resents all package options
}
\examples{
resetAllSechmOptions()

}
