% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{setSechmOption}
\alias{setSechmOption}
\title{setSechmOption}
\usage{
setSechmOption(variable, value)
}
\arguments{
\item{variable}{The name of the variable to set}

\item{value}{The parameter value to save}
}
\value{
None
}
\description{
Sets a package-wide option for `sechm`
}
\examples{
setSechmOption("hmcols", value=c("blue","black","yellow"))

}
